Rename the name and subpath from ai to llm
