+++ llm/DEVELOP.md
@@ -1,4 +1,4 @@
-# ODCAI (OTel based Data Collector for AI)
+# ODCL (OTel based Data Collector for LLM)
 
 **[Semantic Convention](docs/semconv)** |
 **[Support](docs/support/README.md)** |
@@ -7,7 +7,7 @@
 **[License](LICENSE)**
 
 ---
-ODCAI (OTel based Data Collector for AI) is the tool or template to generate OpenTelemetry metrics for various
+ODCL (OTel based Data Collector for LLM) is the tool or template to generate OpenTelemetry metrics for various
 LLM and LLM Applications. All implementation are based on predefined OpenTelemetry Semantic Conventions.
 
 ## Requirements
@@ -24,7 +24,7 @@ java -version
 2) Get the source code from `github.ibm.com`.
 ```bash
 git clone https://github.ibm.com/instana-llmetry/otel-dc.git
-cd otel-dc/ai
+cd otel-dc/llm
 ```
 
 3) Build with Gradle
@@ -33,7 +33,7 @@ cd otel-dc/ai
 ```
 *Note: gradle 7.4 will be installed if you do not have it.*
 
-## Run ODCAI
+## Run ODCL
 
 1) Make sure code is built with Java SDK 11+.
 
@@ -48,8 +48,8 @@ cd otel-dc/ai
 ```
 1) Option-2: Extract deployment package generated by gradle:
 ```bash
-tar vxf otel-dc-ai-*.tar
-cd otel-dc-ai-*
+tar vxf otel-dc-llm-*.tar
+cd otel-dc-llm-*
 ```
 
 Then, user can create the proper configuration files:
@@ -57,4 +57,4 @@ Then, user can create the proper configuration files:
   - config/logging.properties
 
 Run the Data Collector with following command according to your current OS:
-  - bin/odcai
+  - bin/odcl
