+++ llm/architecture.md
@@ -1,14 +1,14 @@
-# ODCAI (OTel based Data Collector for AI)
+# ODCL (OTel based Data Collector for LLM)
 
 ---
-ODCAI (OTel based Data Collector for AI) is the tool or template to generate OpenTelemetry metrics for various
+ODCL (OTel based Data Collector for LLM) is the tool or template to generate OpenTelemetry metrics for various
 LLM and LLM Applications. All implementation are based on predefined OpenTelemetry Semantic Conventions. 
 
 ## Runtime Architecture
 
 ![image](https://media.github.ibm.com/user/104429/files/7e1ba353-477c-448f-aeda-94ee69827586)
 
-## AI Data Collectors
+## LLM Data Collectors
 Watsonx: https://github.ibm.com/instana-llmetry/llm-dc-otel/tree/main/watsonx-dc  
 OpenAI: TBD
