+++ llm/config/logging.properties
@@ -6,7 +6,7 @@ java.util.logging.ConsoleHandler.formatter = java.util.logging.SimpleFormatter
 
 # default file output is in user's home directory.
 java.util.logging.FileHandler.level=INFO
-java.util.logging.FileHandler.pattern = %h/ai-dc%u.log
+java.util.logging.FileHandler.pattern = %h/llm-dc%u.log
 java.util.logging.FileHandler.limit = 500000
 java.util.logging.FileHandler.count = 10
 java.util.logging.FileHandler.formatter = java.util.logging.XMLFormatter
